+++ spec/features/movies/index_spec.rb
@@ -25,6 +25,7 @@ RSpec.describe 'Movie Index Page', type: :feature do
       
       click_link 'Discover Page'
       expect(current_path).to eq(user_discover_path(user1.id))
+      expect(page).to have_content("#{user1.name}'s Discover Page")
     end
   end
 
@@ -33,9 +34,13 @@ RSpec.describe 'Movie Index Page', type: :feature do
       user1 = User.create!(name: 'Andrew', email: 'concertenthusiast@musac.org')
 
       visit user_discover_path(user1.id)
-      fill_in :search, with: 'Home'
+      fill_in :search, with: 'y'
       click_on 'Find Movies'
       expect(current_path).to eq(user_movies_path(user1.id))
+      expect(page).to have_content('Dilwale Dulhania Le Jayenge')
+      expect(page).to have_content('Your Name')
+      expect(page).to have_content('The Good, the Bad and the Ugly')
+      expect(page).to_not have_content('The Green Mile')
     end
   end
 end
